Add dialog before deleting from sidebar
